MacOSX build fixes. Note that it will still not build with -DPOSIX on that platform.

+++ zz_priv.h
@@ -25,7 +25,7 @@
 #define MAX_LEN_PN		(64 * 5 + 1)
 #define MAX_LEN_DATETIME	(26 + 1)
 #define MAX_LEN_CS		(16 + 1)
-#define UNLIMITED		0xffffffff
+#define UNLIMITED		(long)0xffffffff
 
 /// Enumerant for Value Representations. Approach taken from XMedCon.
 enum VR
